+++ app/utils/theme/index.tsx
@@ -1,0 +1,5 @@
+export { default as ThemeWrapper } from "./themeWrapper";
+export { default as palette } from "./palette";
+export * as themes from "./themes";
+export { default as setTheme } from "./setTheme";
+export type { Theme } from "./types";
