+++ app/utils/theme/palette.tsx
@@ -1,0 +1,21 @@
+export const palette = {
+  wave: {
+    800: "#083A40",
+  },
+  sand: {
+    400: "#ECE7D5",
+  },
+  moon: {
+    600: "#0B110F",
+  },
+  passionflower: {
+    500: "#6935ff",
+    100: "#d9d6ff",
+  },
+  lichen: {
+    600: "#8af214",
+  },
+  white: "#fff",
+};
+
+export default palette;
