+++ app/utils/theme/types.tsx
@@ -1,0 +1,16 @@
+export type Theme = {
+  colors: {
+    background: string;
+    accent: string;
+    accent2: string;
+    accent3: string;
+    muted: string;
+    text1: string;
+    text2: string;
+    text3: string;
+  };
+  fonts: {
+    body: string;
+    heading: string;
+  };
+};
